ml-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
aqueduct==1.10.14
asposestorage==1.0.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==3.2.0
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==1.5.3
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
pycparser==2.21
pydantic==1.8
pydantic_core==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
requests-toolbelt==1.0.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.7.4
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
python-dotenv